requirements.txt
Flask
pyjwt
requests
cryptography